mod: modified `unmount` method to avoid a problem where users pass a form ID to the framework.
If a user passes an element ID instead of defining the form using the `form` object, this means, when the `unmount` method is called, the form will be removed from the DOM never to be seen again as we didn't get passed a reference to the DOM structure and at the moment, don't create a virtual DOM to manipulate.

This might be something to change in the future but at the moment, I've disabled being able to unmount forms that weren't created with FormJS. You can still call the method without error, you'll just get a warning in the console.